Add handling for reaction messages from WebSocket
Introduced method to check if the WebSocket message is a reaction and stop its broadcast. A new DTO, ReactionToSocketDTO, was added for reaction data handling and method checkByteArrayForReaction was added in the MessagesController and the NoGuiServer.

This change improves server performance by stopping unnecessary broadcasts, particularly for reaction messages. Reactions are now updated directly in the server and database, reducing traffic and potential lag.

+++ src/main/java/com/soeguet/util/MessageTypes.java
@@ -1,8 +1,8 @@
 package com.soeguet.util;
 
 public interface MessageTypes {
-    int DELETED = 127;
-    int NORMAL = 0;
-    int INTERACTED = 2;
-    int EDITED = 7;
+    byte DELETED = 127;
+    byte NORMAL = 0;
+    byte INTERACTED = 2;
+    byte EDITED = 7;
 }
